Make Licenser plugin work with Gradle 7.1
Gradle 7.1 introduced some changes which caused `project.extensions.create('license', LicenseExtension, project.objects, project)` to return `DefaultBasePluginExtension` (introduced in Gradle 7.1) instead of `LicenseExtension` which would lead to an exception when trying to access a property of `LicenseExtension` on the returned instance.

The issue was traced back to using `project.with { ... }` and not properly resolving the instance variables and field accessors inside that block.

By moving everything out of the `project.with { ... }` block, the field accessor for `Licenser.extension` is properly resolved.